requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.6.12
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.12.0
tensorboard>=2.11.0
opencv-python>=4.6.0
Pillow>=9.0.0
lifelines>=0.27.0
scipy>=1.9.0
tqdm>=4.64.0
